Add toasters

+++ libs/dashboard/view/dashboard-view.component.ts
@@ -1,13 +1,14 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
+import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { combineLatest, map } from 'rxjs';
+import Swal from 'sweetalert2';
 
 import { deleteDashboardRequested, selectDashboard, updateDashboardRequested } from '../../store';
 import { DashboardChartComponent } from '../chart';
 import { Dashboard } from '../dashboard';
-import { Router } from '@angular/router';
 
 @Component({
   standalone: true,
@@ -51,7 +52,20 @@ export class DashboardViewComponent {
     this.router.navigate(['dashboards', dashboard.id, 'edit']);
   }
 
-  onDeleteDashboard(dashboard: Dashboard): void {
-    this.store.dispatch(deleteDashboardRequested({ dashboard_id: dashboard.id }));
+  async onDeleteDashboard(dashboard: Dashboard): Promise<void> {
+    const result = await Swal.fire({
+      title: 'Confirm to delete',
+      text: ' Delete this dashboard forever?',
+      showCancelButton: true,
+      cancelButtonText: 'No, cancel',
+      cancelButtonColor: 'var(bs-secondary)',
+      confirmButtonText: 'Yes, delete it',
+      confirmButtonColor: 'var(--bs-danger)',
+      reverseButtons: true,
+    });
+
+    if (result.isConfirmed) {
+      this.store.dispatch(deleteDashboardRequested({ dashboard_id: dashboard.id }));
+    }
   }
 }
